+++ src/commands/migrate.ts
@@ -1,16 +1,25 @@
+import { execute } from './functions';
+
+let show: string = '';
+
 const migrate = async (vscode: any) => {
-  const terminal = vscode.window.createTerminal({ hideFromUser: true });
-  terminal.sendText('php spark migrate');
+  show = await vscode.window.showQuickPick(['Yes', 'No'], {
+    placeHolder: 'Show in terminal?',
+  });
+
+  if (show.length === 0) {
+    return;
+  }
+
+  execute(vscode, 'migrate', 'php spark migrate', show === 'Yes');
 };
 
 const migrateRefresh = async (vscode: any) => {
-  const terminal = vscode.window.createTerminal({ hideFromUser: true });
-  terminal.sendText('php spark migrate:refresh');
+  execute(vscode, 'migrate refresh', 'php spark migrate:refresh', true);
 };
 
 const migrateRollback = async (vscode: any) => {
-  const terminal = vscode.window.createTerminal({ hideFromUser: true });
-  terminal.sendText('php spark migrate:rollback');
+  execute(vscode, 'migrate rollback', 'php spark migrate:rollback', true);
 };
 
 export { migrate, migrateRefresh, migrateRollback };