+++ src/commands/routes.ts
@@ -1,0 +1,7 @@
+import { execute } from './functions';
+
+const routes = async (vscode: any) => {
+  execute(vscode, 'routes', 'php spark routes', true);
+};
+
+export default routes;